Moving Application layer components out to separate folders.
