+++ Source/Stack/0. Layer 7 - APPLICATION (Transactions, Balances)/Account/AbstractIdentity.swift
