+++ Source/Stack/0. Layer 7 - APPLICATION (Transactions, Balances)/StateSubscriber/DefaultStateSubsciber.swift
@@ -1,0 +1,57 @@
+//
+// MIT License
+//
+// Copyright (c) 2018-2019 Radix DLT ( https://radixdlt.com )
+// 
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+// 
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+// 
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+//
+
+import Foundation
+import RxSwift
+
+public final class DefaultStateSubsciber: StateSubscriber {
+    
+    private let atomStore: AtomStore
+    private let particlesToStateReducer: ParticlesToStateReducer
+    
+    public init(
+        atomStore: AtomStore,
+        particlesToStateReducer: ParticlesToStateReducer = DefaultParticlesToStateReducer.init()
+        ) {
+        self.atomStore = atomStore
+        self.particlesToStateReducer = particlesToStateReducer
+    }
+}
+
+// MARK: StateSubscriber
+public extension DefaultStateSubsciber {
+    func observeState<State>(
+        ofType stateType: State.Type,
+        at address: Address
+    ) -> Observable<State> where State: ApplicationState {
+        
+        return atomStore.onSync(address: address)
+            .map { [unowned self] date in
+                let upParticles = self.atomStore.upParticles(at: address, stagedUuid: nil)
+                let reducedState = try self.particlesToStateReducer.reduce(upParticles: upParticles, to: stateType)
+                return reducedState
+        }
+    }
+}
+
