+++ Source/Stack/0. Layer 7 - APPLICATION (Transactions, Balances)/TransactionMaker/ResultOfUserAction/SubmitAtomError.swift
