+++ Source/Stack/0. Layer 7 - APPLICATION (Transactions, Balances)/TransactionMaker/TransactionMaker.swift
@@ -1,0 +1,46 @@
+//
+// MIT License
+//
+// Copyright (c) 2018-2019 Radix DLT ( https://radixdlt.com )
+// 
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+// 
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+// 
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+//
+
+import Foundation
+import RxSwift
+
+public protocol TransactionMaker: TransactionToAtomMapper {
+    func send(transaction: Transaction, toOriginNode: Node?) -> ResultOfUserAction
+}
+
+public extension TransactionMaker {
+    
+    func send(transaction: Transaction) -> ResultOfUserAction {
+        return send(transaction: transaction, toOriginNode: nil)
+    }
+    
+    func execute(actions: [UserAction], originNode: Node? = nil) -> ResultOfUserAction {
+        let transaction = Transaction { actions }
+        return send(transaction: transaction, toOriginNode: originNode)
+    }
+    
+    func execute(actions: UserAction..., originNode: Node? = nil) -> ResultOfUserAction {
+        return execute(actions: actions, originNode: originNode)
+    }
+}
